+++ resources/lang/de/accounting.php
@@ -67,10 +67,11 @@ return [
     'no_cost_center' => 'Keine Kostenstelle',
     'invalid_cost_center_x' => 'Ungültige Kostenstelle \':cost_center\'',
     'use_cost_centers' => 'Kostenstellen verwenden',
-    'controlling' => 'Controlling',
     'controlled' => 'Kontrolliert',
     'mark_controlled' => 'Als kontrolliert markieren',
     'controlled_at' => 'Kontrolliert am',
     'controlled_by' => 'Kontrolliert von',
-
+    'cannot_update_already_controlled_transaction' => 'Cannot update already controlled transaction.',
+    'already_controlled' => 'Already controlled.',
+    'controlling_by_user_who_created_not_allowed' => 'The user who registered the transaction is not allowed to mark it as controlled.',
 ];