+++ resources/lang/en/accounting.php
@@ -67,9 +67,11 @@ return [
     'no_cost_center' => 'No cost center',
     'invalid_cost_center_x' => 'Invalid cost center \':cost_center\'',
     'use_cost_centers' => 'Use cost centers',
-    'controlling' => 'Controlling',
     'controlled' => 'Controlled',
     'mark_controlled' => 'Mark as controlled',
     'controlled_at' => 'Controlled at',
     'controlled_by' => 'Controlled by',
+    'cannot_update_already_controlled_transaction' => 'Kann bereits kontrollierte Transaktion nicht ändern.',
+    'already_controlled' => 'Bereits kontrolliert.',
+    'controlling_by_user_who_created_not_allowed' => 'Der Benutzer welche eine Transaktion registriert hat ist nicht ermächtigt sie auch als kontrolliert zu markieren.',
 ];